add full state dict saving and testing
